Updating the template.

+++ game/main_scene.cpp
@@ -1,15 +1,5 @@
 #include "game/main_scene.h"
 
-#include "input/input.h"
-#include "math/random.h"
-#include "object/object_allocator.h"
-#include "os/thread.h"
-#include "os/semaphore.h"
-#include "os/mutex.h"
-#include "physics/physics_2d.h"
-#include "scene/scene_manager.h"
-
-
 void MainScene::init(const CreateInfo&)
 {
     mark(MARK_UPDATE);
@@ -24,6 +14,6 @@ void MainScene::exit()
 {
 }
 
-void MainScene::update(f64)
+void MainScene::update(f32)
 {
 }
